Added Compatibility to UI7 and unlimited listing
We have added compatibility to UI7 and now the We can list unlimited
items.

This version is very unstable and the thermostat does not work very
well. There is a bug in UI7 while requesting sdata where the  heat set
point value shown is the same as the cool value which is wrong so in
order to get the real heat value We use the following:

http://veraip:3480/data_request?id=variableget&DeviceNum=69&serviceId=ur
n:upnp-org:serviceId:TemperatureSetpoint1_Heat&Variable=CurrentSetpoint

The addon is very unstable and does not exit properly.